Creating isLoggedIn middleware and Make the Dashboard, Sidebar of the UI

+++ Backend/seeder.js
@@ -1,4 +1,3 @@
-import mongoose from "mongoose";
 import dotenv from "dotenv";
 import connectDB from "./config/db.js";
 import products from "./data/products.js";
@@ -9,12 +8,16 @@ connectDB();
 
 const importData = async () => {
   try {
-    await Product.deleteMany();
-    await Product.insertMany(products);
-    console.log("✅ Products seeded successfully");
-    process.exit();
+    if(process.env.NODE_ENV = "develeopment"){
+      await Product.deleteMany();
+      await Product.insertMany(products);
+      console.log("✅ Products seeded successfully");
+      process.exit();
+    } else {
+      console.log("Seeding blocked , not in development mode");
+    }
   } catch (error) {
-    console.log("❌ Seeding failed");
+    console.log("❌ Seeding failed", error);
     process.exit(1);
   }
 };
